+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Woocommerce Transaction Tracking 1.2.0
+
+- Fix issue with not waiting for WP head
+- Update error notice
+
 # Woocommerce Transaction Tracking 1.1.0
 
 - Add Woocommerce checking and admin notice